+++ packages/core/icon/glyph/home-circle.js
@@ -15,7 +15,7 @@ function _extends() { _extends = Object.assign || function (target) { for (var i
 
 var HomeCircleIcon = function HomeCircleIcon(props) {
   return _react.default.createElement(_Icon.default, _extends({
-    dangerouslySetGlyph: "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\" focusable=\"false\" role=\"presentation\"><g fill-rule=\"evenodd\"><path d=\"M12 22c5.523 0 10-4.477 10-10S17.523 2 12 2 2 6.477 2 12s4.477 10 10 10z\" fill=\"currentColor\"/><path d=\"M11.643 6.357a.506.506 0 0 1 .714 0l4.931 4.931c.393.393.256.712-.29.712H7.002c-.553 0-.68-.323-.29-.712l4.93-4.93z\" fill=\"inherit\"/><path d=\"M8 12.003c0-.554.453-1.003.997-1.003h6.006c.55 0 .997.438.997 1.003v4.994c0 .554-.453 1.003-.997 1.003H8.997A.993.993 0 0 1 8 16.997v-4.994zm2.667 3.514V18h2.666v-2.483a1.334 1.334 0 1 0-2.666 0z\" fill=\"inherit\"/></g></svg>"
+    dangerouslySetGlyph: "<svg width=\"24\" height=\"24\" viewBox=\"0 0 24 24\" focusable=\"false\" role=\"presentation\"><path d=\"M10.5 16H9a1 1 0 0 1-1-1v-4.946l3.28-3.379.021-.021a1 1 0 0 1 1.414.021L16 10.059V15a1 1 0 0 1-1 1h-1.5v-2.5a1.5 1.5 0 0 0-3 0V16zm7.5-3.936a1 1 0 0 0 1.215-1.564L14.15 5.282a3 3 0 0 0-4.242-.063l-.063.063L4.78 10.5A1 1 0 0 0 6 12.061V15a3 3 0 0 0 3 3h6a3 3 0 0 0 3-3v-2.936zM12 22C6.477 22 2 17.523 2 12S6.477 2 12 2s10 4.477 10 10-4.477 10-10 10z\" fill=\"currentColor\" fill-rule=\"evenodd\"/></svg>"
   }, props));
 };