+++ packages/editor/editor-core/src/plugins/media/pm-plugins/main.ts
@@ -78,8 +78,6 @@ export class MediaPluginState {
 
   public pickers: PickerFacade[] = [];
   private popupPicker?: PickerFacade;
-  // @ts-ignore
-  private clipboardPicker?: PickerFacade;
   private dropzonePicker?: PickerFacade;
   // @ts-ignore
   private customPicker?: PickerFacade;
@@ -514,7 +512,6 @@ export class MediaPluginState {
     pickers.splice(0, pickers.length);
 
     this.popupPicker = undefined;
-    this.clipboardPicker = undefined;
     this.dropzonePicker = undefined;
     this.customPicker = undefined;
   };
@@ -561,14 +558,6 @@ export class MediaPluginState {
         );
 
         pickers.push(
-          (this.clipboardPicker = await new Picker(
-            'clipboard',
-            pickerFacadeConfig,
-            defaultPickerConfig,
-          ).init()),
-        );
-
-        pickers.push(
           (this.dropzonePicker = await new Picker(
             'dropzone',
             pickerFacadeConfig,
@@ -596,7 +585,7 @@ export class MediaPluginState {
     pickers.forEach(picker => picker.setUploadParams(uploadParams));
   }
 
-  private trackNewMediaEvent(pickerType: string) {
+  public trackNewMediaEvent(pickerType: string) {
     return (mediaState: MediaState) => {
       analyticsService.trackEvent(
         `atlassian.editor.media.file.${pickerType}`,