+++ packages/media/media-ui/src/BlockCard/ResolvingView/index.tsx
@@ -3,6 +3,8 @@ import Spinner from '@atlaskit/spinner';
 import { CollapsedFrame } from '../CollapsedFrame';
 import { minWidth, maxWidth } from '../dimensions';
 import { SingleLineLayout } from '../SingleLineLayout';
+import { FormattedMessage } from 'react-intl';
+import { messages } from '../../messages';
 
 export interface BlockCardResolvingViewProps {
   /** The optional click handler */
@@ -23,7 +25,10 @@ export class BlockCardResolvingView extends React.Component<
         maxWidth={maxWidth}
         onClick={onClick}
       >
-        <SingleLineLayout left={<Spinner size="small" />} middle="Loading..." />
+        <SingleLineLayout
+          left={<Spinner size="small" />}
+          middle={<FormattedMessage {...messages.loading} />}
+        />
       </CollapsedFrame>
     );
   }