+++ packages/media/media-ui/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-ui
 
+## 11.2.0
+- [minor] [061eb57bab](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/061eb57bab):
+
+  - Added partial i18n support to media-ui
+
 ## 11.1.2
 - [patch] [6a52b3d258](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6a52b3d258):